wr-265: * move constants to the folder
